+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Change Log
 * Raname Report -> CoverallsReport
 * Raname AdaptorInterface -> EnviromentAdaptor
 * Raname RecordInterface -> FileRecord
+* Raname ReportParserInterface -> ReportParser
 
 
 1.4.0.3